refactor: split query plane helpers for readability
